.claude/skills/starter-django/assets/templates/backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4
django-cors-headers>=4.6,<5
dj-database-url>=2.3,<3
gunicorn>=23,<24
whitenoise>=6.8,<7
psycopg[binary]>=3.2,<4
PyJWT[crypto]>=2.11,<3
python-decouple>=3.8,<4
